// web/src/api/models/collection-config.ts
export const CollectionConfigCreationModeEnum = {
    aperag_advanced: 'aperag_advanced',
    mirofish_simple: 'mirofish_simple'
} as const;

export type CollectionConfigCreationModeEnum = typeof CollectionConfigCreationModeEnum[keyof typeof CollectionConfigCreationModeEnum];

export const CollectionConfigGraphEngineEnum = {
    lightrag: 'lightrag',
    mirofish: 'mirofish'
} as const;

export type CollectionConfigGraphEngineEnum = typeof CollectionConfigGraphEngineEnum[keyof typeof CollectionConfigGraphEngineEnum];